# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.Task:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      is_completed:
        type: boolean
      kanban_id:
        type: integer
      title:
        type: string
      updated_at:
        type: string
    type: object
...
  response.JSONResponse:
    description: default JSON-response
    properties:
      data:
        type: object
      error:
        type: string
      success:
        type: boolean
    type: object
  task_handler.CreateTaskRequestParam:
    properties:
      description:
        type: string
      is_completed:
        type: boolean
      kanban_id:
        description: |-
          Add other necessary fields for creation here, e.g., ProjectID, KanbanID, etc.
          Assuming these might come from the route or other means if not in the body
        type: integer
      title:
        type: string
    required:
    - kanban_id
    - title
    type: object
  task_handler.UpdateTaskRequestParam:
    properties:
      description:
        type: string
      is_completed:
        description: Use pointer to distinguish between false and not provided
        type: boolean
      title:
        type: string
    type: object
...
info:
  contact: {}
  description: Task manager
  title: DataTask
  version: "1.0"
paths:
...
  /kanban_tasks/{kanban_id}:
    get:
      description: Get all tasks associated with a Kanban board
      parameters:
      - description: Kanban Board ID
        in: path
        name: kanban_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.JSONResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.Task'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.JSONResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.JSONResponse'
      summary: Get Tasks by Kanban ID
      tags:
      - Task
...
  /project_tasks/{project_id}:
    get:
      description: Get all tasks associated with a project
      parameters:
      - description: Project ID
        in: path
        name: project_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.JSONResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.Task'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.JSONResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.JSONResponse'
      summary: Get Tasks by Project ID
      tags:
      - Task
...
  /task:
    post:
      consumes:
      - application/json
      description: Create a new Task
      parameters:
      - description: Task data for creation
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/task_handler.CreateTaskRequestParam'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/response.JSONResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.Task'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.JSONResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.JSONResponse'
      summary: Create Task
      tags:
      - Task
  /task/{id}:
    delete:
      description: Delete a Task by its ID
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            $ref: '#/definitions/response.JSONResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.JSONResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.JSONResponse'
      summary: Delete Task
      tags:
      - Task
    get:
      description: Get a Task by its ID
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.JSONResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.Task'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.JSONResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.JSONResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.JSONResponse'
      summary: Get Task by ID
      tags:
      - Task
    put:
      consumes:
      - application/json
      description: Update a Task's details
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated Task data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/task_handler.UpdateTaskRequestParam'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.JSONResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.Task'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.JSONResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.JSONResponse'
      summary: Update Task
      tags:
      - Task
  /task/{task_id}/assign:
    post:
      consumes:
      - application/json
      description: Assign a user to a task
      parameters:
      - description: Task ID
        in: path
        name: task_id
        required: true
        type: integer
      - description: User ID to assign
        in: body
        name: request
        required: true
        schema:
          additionalProperties:
            type: integer
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.JSONResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.JSONResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.JSONResponse'
      summary: Assign User to Task
      tags:
      - Task
  /user/{user_id}/tasks:
    get:
      description: Get all tasks assigned to a user
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.JSONResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.Task'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.JSONResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.JSONResponse'
      summary: Get Tasks by User ID
      tags:
      - Task
...
security:
- Authorization: []
securityDefinitions:
  Authorization:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"